% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fs_server}
\alias{fs_server}
\title{Fullscreen Server}
\usage{
fs_server(target = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{target}{Id of target element to make full screen,
if \code{NULL} makes the entire page full screen.}

\item{session}{A valid shiny session.}
}
\description{
Trigger a full screen from the server.
}
\examples{
library(shiny)

ui <- fluidPage(
	h1("Hello fullscreen"),
	actionButton(
		"fs",
		"Fullscreen via server"
	)
)

server <- function(input, output, session){
	observeEvent(input$fs,{
		fs_server()
	})
}

if(interactive())
 shinyApp(ui, server)

}
