% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fsTrigger}
\alias{fsTrigger}
\alias{fsButton}
\alias{fsLink}
\title{Fullscreen Triggers}
\usage{
fsButton(
  ...,
  id = NULL,
  class = NULL,
  .target = NULL,
  .type = "default",
  .tag = htmltools::a
)

fsLink(..., id = NULL, class = NULL, .target = NULL, .tag = htmltools::a)
}
\arguments{
\item{...}{Arguments to pass to the button \code{.tag}.}

\item{id}{Id of the button.}

\item{class}{Class to assign to the button.}

\item{.target}{Id of target element to make full screen,
if \code{NULL} makes the entire page full screen.}

\item{.type}{Button type, e.g.: \code{success}.}

\item{.tag}{Htmltools or shiny tag to use to create the button.
Default to \verb{<a>}.}
}
\description{
Create a button or a link that triggers the fullscreen.
}
\examples{
library(shiny)

ui <- fluidPage(
	h1("Hello fullscreen"),
	fsButton("Make fullscreen")
)

server <- function(...){}

if(interactive())
 shinyApp(ui, server)

}
