% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{make_script}
\alias{make_script}
\title{Create Script for Button}
\usage{
make_script(id, target)
}
\arguments{
\item{id}{Id of button.}

\item{target}{Target element}
}
\description{
Create Script for Button
}
\keyword{internal}
